# docs/api/openApiSpecs.json
{
  "info": {
    "title": "EverLearn: e-learning platform",
    "version": "1.0.1",
    "description": "API documentation"
  },
  "host": "localhost:8000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "paths": {
...
    "/blockStudents/courses/{courseId}": {
      "get": {
        "tags": [
          "blockStudent"
        ],
        "summary": "To get all blocked students of a specific course",
        "responses": {
          "200": {
            "description": "successful operation.",
            "schema": {
              "$ref": "components/blockStudents/res.json#GetBlockStudentCourse"
            }
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      }
    },
    "/courses": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "To get list of courses.",
        "responses": {
          "200": {
            "description": "successful operation.",
            "schema": {
              "$ref": "components/courses/res.json#GetCourses"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "To create courses.",
        "parameters": [
          {
            "in": "body",
            "name": "courses",
            "description": "courses data",
            "schema": {
              "$ref": "components/courses/req.json#CreateCourse"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation."
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      }
    },
    "/courses/{courseId}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "To get courses.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "courseId",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation.",
            "schema": {
              "$ref": "components/courses/res.json#GetCourse"
            }
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "courses"
        ],
        "summary": "To update courses.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "courseId",
            "type": "integer"
          },
          {
            "in": "body",
            "name": "courses",
            "description": "courses data",
            "schema": {
              "$ref": "components/courses/req.json#UpdateCourse"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation."
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "courses"
        ],
        "summary": "To delete courses.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "courseId",
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation."
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      }
    },
    "/courses/{courseId}/ratings": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "To get list of courses ratings.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "courseId",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation.",
            "schema": {
              "$ref": "components/courses/res.json#GetCourseRatings"
            }
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "To create courses.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "courseId",
            "type": "integer"
          },
          {
            "in": "body",
            "name": "courses rating",
            "description": "courses rating data",
            "schema": {
              "$ref": "components/courses/req.json#AddCourseRating"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation."
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      }
    },
    "/courses/{courseId}/modules": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "To get list modules of course.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "courseId",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation.",
            "schema": {
              "$ref": "components/courses/res.json#GetCourseModules"
            }
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "To create course module.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "courseId",
            "type": "integer"
          },
          {
            "in": "body",
            "name": "courses module",
            "description": "courses module data",
            "schema": {
              "$ref": "components/courses/req.json#CreateCourseModule"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation."
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/enrollments/courses/{courseId}": {
      "get": {
        "tags": [
          "enrollments"
        ],
        "summary": "To get specific course enrollment.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "courseId",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation.",
            "schema": {
              "$ref": "components/enrollment/res.json#GetEnrollmentsCourse"
            }
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
...
  "securityDefinitions": {},
...
}